inline all custom element define stuff
