--- packages/independent/custom-elements-redefine/src/internal/custom_element_class_map.js
@@ -1,1 +1,0 @@
-export const customElementClassMap = new Map();
